remove old cli; move main.py to top-level
