+++ src/model_metadata/__main__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from model_metadata.main import mmd
+
+if __name__ == "__main__":
+    raise SystemExit(mmd())
